requirements.txt
Flask
gunicorn
boto3
botocore
python-dotenv
requests
psutil